+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2021-06-13
+
+### Changed
+
+- Update metadata Polkadot runtime/9220
+- Update metadata Kusama runtime/9230
+- Update metadata Westend runtime/9230
+
 ## [0.3.0] - 2021-03-17
 
 ### Changed